Fix popular manga

+++ Models/PopularMangaResponse.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace AkariApi.Models
 {
     public class PopularMangaResponse
@@ -13,7 +11,7 @@ namespace AkariApi.Models
         public required string Type { get; set; }
         public required string[] Authors { get; set; }
         public required string[] Genres { get; set; }
-        public required int Score { get; set; }
+        public required decimal Score { get; set; }
         public int? MalId { get; set; }
         public int? AniId { get; set; }
         public DateTimeOffset CreatedAt { get; set; }
